DynamicArray test rm over capacity error test.
Many code sanitizers have lots of trouble with dynamic_array_reserve
with a impossible large size. Although the test only tries to see
whether we catch it alright. It presents major challenges when
-fsanitize=address is provided to a compiler.
Valgrind is afraid one is mallocing a with a negative number,
so all in all I would like to remove it.